spent 3/4 of a day rewriting a for loop
had my mind set on wrapping the whole NN training loop in a jit region.
the problem before was that while we can evaluate loss and gradient
quickly with jitted functions, they were re-jitted at the start of each
NN training round, which made the overall solver slow.

Now we jit the whole training loop once, and it is very fast after that.
The loop is reformulated with jax.lax.scan, a very nice way but had to
wrap my head around it.

obvs introduced quite a few breaking changes. the nn_wrapper class now
does almost nothing in its constructor, most of the business logic is in
the training jit region now (so is compiled away pretty much). it could
be argued the more 'jax-y' way to do it would be without the wrapper
class entirely.