new attempt at sampler: ~~~ geometric mcmc ~~~
geometric bc. it is supposed to alleviate the problem of weird geometry
in terminal condition space. basically we try to do mcmc in x(0), but
instead of directly taking a step, we estimate a suitable step in
terminal conditions (via linearization) which modifies our proposal
distribution, but M-H correction should make everything nice again.