experiment with sampling from value sublevel set.
basically the S = {x | V(x) <= V_max} i wrote in the paper without
thinking too much about it.

necessitated some changes to the mala sampler transition info function.
specifically, we cannot just evaluate the descent direction
-dlogpdf/dx0, instead we have to take into account that logpdf also
depends on v0, which depends on λT, which depends on x0 also (the last
one due to assuming PMP is invertible).

coded that up, works alright, not sure if it is a huge improvement.
it might turn out that it improves the geometry of the preimage of S
under PMP, it looks that way at first glance.

but also hard to debug because we cannot directly plot reward(x0) because
we dont know v0 for all x. (could be plotted when we have a fit for it
though).

at the moment i am putting this away and working to get some results for
the other case.