more plots!
it looks like there is something slightly wrong with the generated data
that causes higher approx error in λ1. the error is mostly really small
but on some vertical bands there are outliers. could be ode integrator
error weirdness or similar. probably okay to just ignore this for now as
it is a quite small amount of data. could probs be fixed by running the
sampler again with slightly different parameters, like smaller dt or
float64 instead of 32.

next up: make closed loop evaluation!