tried a dumb idea
basically, return the proposed samples as ouptut, as opposed to the
accepted ones. hoped for a similar-ish distribution, a bit smoothed
(bc. we are basically adding a proposal dist to the stationary dist)
but most importantly without long repeated streaks.

but i underestimated how far out of the interesting set most proposals
are. this worked quite bad. putting it in this branch to not disturb the
rest.