first DDP forward pass & lots of plotting!!!
initial (optimal, from backward integration) trajectory is replicated
nicely up to small deviations. for slightly (N(0, 0.1I)) perturbed x0 it
seems that all looks good too. this might be a success!

in rare cases the S matrix briefly stops being p.d.. is this possible in
theory? is it numerical error (looks a bit large for that though)? wrong
implementation? I seem to recall that in one of the papers I read it was
specifically mentioned as an upside of the continuous time version that
the value hessian stays p.d....