bit of a refactor.
had the goal of doing the backward pass the "proper" way without
differentiating the forward solution's interpolation. to do that, we
need the costate that produced the last forward pass, and for that, we
need the forward and backward pass before that.

to make it nicer (scan iteration is markovian) i switched the order of
forward and backward pass. longer explanation in idea dump.

now it works about as well as before at first glance but that paricular
source of error is gone :)