see what happens...
when we try to get a distribution of xf by forward simulating
with approximately optimal controller (LQR here) to then initialise
the backward simulation.

we see the old challenges. the fastest pole limits everything due to
backward instability. after a couple time constants accuracy is out the
window (that is, which trajectory precisely do we get, not necessarily
that the v/vx/vxx on this trajectory are inaccurate...)