debugging \o/
had issue with chattering of u* near active set switching. found
unsuitable use of floating point math, improved it, chattering now much
less.

it seems the main pitfall now are non-pd vxx. is this a bug or a
feature? they are non-pd for a long time, but just before it all goes
wrong one eigenvalue tends to go very low...