sobolev training works!
well, it runs without producing errors or NaNs. but isn't that a
good point to call it a day?!? (but first loss plot looks quite nice)

hessian approximation with random Hvp. Essentially corresponds to
calculating only one column of the hessian.

had to fight jax a bit but got around to it. now we are using the exact
same data format as given by the ode solver, a vmapped dict with entries
['x', 'v', 'vx', 'vxx'].

was not yet able to figure out how to include or exclude calculating
the vxx loss and gradient at (jit) compile time. should be possible
though with some static_argnums or lexical closure magic.

next: throw away the old code :)))