cost function change
to make state space transformation more correct. if we want to learn the
value function etc in terms of z = T(x) (e.g. T(phi) = sin phi, cos phi)
then l(x, u) should probably be a proper function of z. the version
before penalised for example Phi = 2 pi heavily, but it should be the
same as Phi = 0. This is fixed now -- basically we have a quadratic cost
in terms of T(x).

still thinking about how to implement this transformation elsewhere...