terminate solver if ||vxx|| too high
this is a fundamental feature of (some) solutions, not a numerical
fluke.

intuitively, if ||vxx|| goes to infinity (fininte escape time)
then we should arrive at a point where many trajectories arrive at a
focal point (in backward time), thus we can stop integrating them
anyway. will have to find a concrete proof for this though.