coordinate transformation experiments.
this is still unfinished, not sure how to handle vx and esp. vxx
correctly, and also not sure how to implement it software engineering
wise. do we first transform then normalise? if so, we need to account
for the normalisation too when putting it all in the loss function. do
we normalise first then transform? this probably makes the least sense
as then the entire manifold will be different. do we make a unified
transform+normalise map? maybe this is the easiest.

do we autodiff everything in the loss function? ie just say
  loss = || nabla_x V_nn(T(x)) - v_x || + || hess_x V_nn(T(x)) dir - v_xx dir ||
(with dir our nice random vector for the hessian approx)? this is
probably the cleanest but will require the nn to know about T which atm
it doesn't.

the other option is to precalculate all these transformations + jacbians
+ whatever terms we need for the hessian, and then we write a loss
equivalent to the above.

probably this is not worth it though -- the compuatational overhead of
recalculating T and derivatives every time is probably tiny.