tangent/normal projection in sobolev loss :)
the costate (both training data and NN output) is now only evaluated in
the "correct" directions in the loss function. Consider this:

 loss = || (vx_train - vx_nn) P_TxM v ||_2

this would be a loss penalising the error in the directional derivative
in a particular direction P_Txm v, which due to the projection is a
member of the tangent space. we may also regard everything left of v a
member of the cotangent space. we remove the v from the equation and
penalise the whole linear operator, essentially "undoing" the classic
stochastic approximation trick from Czarnecki, 2017.