debugging session
finding out why the NN fits are bad. it was due to data normalisation,
must be some error there. now we have extra plots. and also explained
that the weird loop-shaped thing in the v_mean, v_std plot is the image
of {0} x M, where due to the origin-focussed sampling the data is
concentrated. pretty neat.