be more lax about the value level set
more precisely: previously it happened regularly that a single point
above σ_max threw the estimation of v_k off, i.e. there is a much larger
value level set in which we have *mostly* sigma low enough but not
always.

this is now implemented in the v_k estimator: we find the largest level
set such that a given (large) fraction of points inside of it is below
σ_max.

also, shuffled around params & made categories for them.
also, add noise to value prior to (maybe?) increase sigma. not sure why.
also, rewrote that comment about orthonormal basis for normal space.