remove data normalisation!!!!
the new loss which is scaled by label magnitude pointwise makes v-axis
normalisation obsolete. the x axis normalisaiton was turned off anyway
because of complications with the tangent space projection (i think).
we just need to rescale to different state space units so they are all
close-ish to unity. but no need to be very particular about that as far
as i can tell.