messy 3/4 day of hacking
mostly making my head smoke about the pruning strategies. how to do
this? made implementation of "value substeps" type thing, not sure if
works. noticed that we might just as well do smaller value steps in the
first place and not introduce a second level of looping.

in other news:
- plotting function to plot value along random lines (projected to M)
  between two points close to the current level set
- finally implemented that idea about pruning "the rest" of a trajectory
  before a known suboptimal point. not possible to properly test unless
  the whole pruning is somewhat implemented.
- move simpler type of data thinning where it belongs
- ensure at least 2000 NN iters, helps it at start sometimes
- linear x scale for training plots