euler launcher, argument parser
entry point is launcher.py. there we can specify the argument
permutations with a simple dict of lists of arguments.

it generates commands with the arguments as command line flags. the
flags are parsed in flatquad_landing_experiment.py -- every algo_params
entry of type (bool, int, float, str) can be set with the launcher. if
not given the value from the original algo_params dict is used. very
nice :)

also, stashing old code away