lipschitz pruning it is!
could not stop fantasizing about that perfect capturing of
nonsmoothness, different local solutions, mixture of experts, even found
some previous papers about almost the same thing! but ultimately again i
have to admit it is too late to start stuff like this, so i am deciding
to go with smooth approximation despite the inherently nonsmooth
solution.

and if we admit we only want a smooth solution we might as well just
specify lipschitz constants for the function and its gradient (and hope
it works somewhat decently). this seems nicer to me than somehow pruning
the data based on the "smoothness priors" hidden in the NN size,
activation function, training procedure.

and in the end we get to show some interesting failure cases and propose
further research!

more concretely:
 - first running implementation of lipschitz pruning
   a) prune based on V lipschitz cone-shaped upper bound on V
   b) prune based on Vx lipschitz quadratic upper bound on V
   c) among the remaining, remove any conflicts wrt Vx lipschitz
 - fat euler job dump
 - removed some old code