bit of algoparam tuning
orbits example works now +/- as it should :))) also added option to
inclde "future" (= physical past) data that we already have within the
(presumed) next value band. does not seem to change the behaviour that
much here though. maybe because it is fundamentally not a good idea, but
maybe also just due to low-dim effects, causing the proposals to pretty
much always be about uniform due to "max_kernel" sampling. we'll see how
it does on the 6D example.