2 small tweaks
a) option to scale down kernel so its max is <1

the idea is that a kernel with a max of 1 (and a relatively large
radius) makes the sampling "deteriorate" to a uniform distribution
quickly, removing possible benefits of sampling where uncertainty is
high. with this scaling we assume that adding data reduces uncertainty
not to 0 but decreases it by some fraction, which is probably closer to
the reality of NN training.

b) shorter second training run

no brainer. large learning rates for already almost trained NN make no
sense. purpose of the second run is basically only to let weight decay
pull on the unnecessary weights that formed due to linear huber gradient
from outliers.