different initial guess
turns out if we copy the inputs of a continuous time solution, and apply
them with ZOH and discretised dynamics, things don't always turn out
good.

instead now we get the initial guess by "abusing" the ct controller in
dt. but that is a small price for gaining feedback wrt open loop.